requirements.txt
anyio==3.5.0
asgiref==3.5.0
astroid==2.11.2
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
click==8.1.2
cryptography==36.0.2
dill==0.3.4
dnspython==2.2.1
ecdsa==0.17.0
fastapi==0.75.1
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
motor==2.5.1
passlib==1.7.4
platformdirs==2.5.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pylint==2.13.4
pymongo==3.12.3
python-dotenv==0.20.0
python-jose==3.3.0
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.1
typed-ast==1.5.2
typing_extensions==4.1.1
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.2
wrapt==1.14.0
zipp==3.8.0
